backend/requirements.txt
# Azure Functions O&G RAG API - Lightweight (OpenAI embeddings)

azure-functions>=1.17.0
pinecone>=3.0.0
openai>=1.0.0
anthropic>=0.39.0
